Added check for XDebug mode as an additional requirement to enable FrankenPHP worker mode to avoid performance issues.
See the discussion with Kévin Dunglas in the PR discussion for the reasoning: #932

Essentially, disabling worker mode comes with performance drawbacks, but it's still the only sensible way to use FrankenPHP with XDebug.
By automatically turning off worker mode if
 a) the amount of workers isn't set explicitly,
 b) we're running locally, and
 c) XDebug is not installed, or it's mode is set to "off",
we get both optimal performance and XDebug working out of the box.